+++ src/main/java/joshie/harvest/npcs/item/ItemNPCSpawner.java
@@ -1,13 +1,14 @@
 package joshie.harvest.npcs.item;
 
+import joshie.harvest.api.npc.NPC;
 import joshie.harvest.core.HFTab;
 import joshie.harvest.core.base.item.ItemHFFML;
 import joshie.harvest.npcs.HFNPCs;
-import joshie.harvest.api.npc.NPC;
 import joshie.harvest.npcs.NPCHelper;
 import joshie.harvest.npcs.entity.EntityNPC;
 import joshie.harvest.town.TownHelper;
 import net.minecraft.client.renderer.block.model.ModelResourceLocation;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
@@ -17,9 +18,13 @@ import net.minecraft.util.EnumHand;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import net.minecraft.world.WorldServer;
 import net.minecraftforge.client.model.ModelLoader;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
+
+import javax.annotation.Nonnull;
+import java.util.UUID;
 
 import static joshie.harvest.core.lib.HFModInfo.MODID;
 
@@ -38,19 +43,32 @@ public class ItemNPCSpawner extends ItemHFFML<ItemNPCSpawner, NPC> {
         return getObjectFromStack(stack).getLocalizedName();
     }
 
+    private void spawnNPC(World world, BlockPos pos, NPC npc) {
+        EntityNPC entity = NPCHelper.getEntityForNPC(world, npc);
+        entity.setPosition(pos.getX() + 0.5, pos.getY() + 1, pos.getZ() + 0.5);
+        entity.resetSpawnHome();
+        if (npc == HFNPCs.CARPENTER) {
+            entity.setUniqueId(TownHelper.getClosestTownToEntity(entity).getID());
+        }
+
+        world.spawnEntityInWorld(entity);
+    }
+
     @Override
+    @Nonnull
     public EnumActionResult onItemUse(ItemStack stack, EntityPlayer player, World world, BlockPos pos, EnumHand hand, EnumFacing facing, float hitX, float hitY, float hitZ) {
         NPC npc = getObjectFromStack(stack);
         if (npc != null) {
             if (!world.isRemote) {
-                EntityNPC entity = NPCHelper.getEntityForNPC(world, npc);
-                entity.setPosition(pos.getX() + 0.5, pos.getY() + 1, pos.getZ() + 0.5);
-                entity.resetSpawnHome();
                 if (npc == HFNPCs.CARPENTER) {
-                    entity.setUniqueId(TownHelper.getClosestTownToEntity(entity).getID());
-                }
-
-                world.spawnEntityInWorld(entity);
+                    WorldServer server = (WorldServer) world;
+                    UUID uuid = TownHelper.getClosestTownToEntity(player).getID();
+                    Entity entity = server.getEntityFromUuid(uuid);
+                    if (entity instanceof EntityNPC) {
+                        entity.setPosition(pos.getX() + 0.5, pos.getY() + 1, pos.getZ() + 0.5);
+                        ((EntityNPC)entity).resetSpawnHome();
+                    } else spawnNPC(world, pos, npc);
+                } else spawnNPC(world, pos, npc);
             }
 
             stack.splitStack(1);